+++ app/Request.php
@@ -3,6 +3,9 @@
 use Illuminate\Database\Eloquent\Model;
 use Carbon\Carbon;
 
+use App\Request as MentorshipRequest;
+use App\Status;
+
 class Request extends Model
 {
 
@@ -119,4 +122,22 @@ class Request extends Model
             }
         });
     }
+
+    /**
+     * Get all OPEN requests that have been made in
+     * the last 24 hours or more
+     *
+     * @return array
+     */
+    public static function getUnmatchedRequests($duration = 24)
+    {
+        $threshold_date = Carbon::now()->subHours($duration);
+        $unmatched_requests = MentorshipRequest::with('user')
+            ->where('status_id', Status::OPEN)
+            ->whereDate('created_at', '<=', $threshold_date)
+            ->get();
+
+        return $unmatched_requests;
+    }
+
 }