+++ config/notifications.php
@@ -1,0 +1,18 @@
+<?php
+
+return [
+    "local" => [
+        "placed_unmatched_fellows_notification_email" =>
+            "test-user-admin@andela.com"
+    ],
+
+    "staging" => [
+        "placed_unmatched_fellows_notification_email" =>
+            "test-user-admin@andela.com"
+    ],
+
+    "production" => [
+        "placed_unmatched_fellows_notification_email" =>
+            "mentorshiprequest@andela.com"
+    ]
+];